backend/pyproject.toml
[project]
name = "obsidian-wiki-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "sqlalchemy[asyncio]>=2,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "aiofiles>=24,<25",
    "redis[hiredis]>=5,<6",
    "python-jose[cryptography]>=3.3,<4",
    "bcrypt>=4,<5",
    "gitpython>=3.1,<4",
    "python-frontmatter>=1,<2",
    "python-multipart>=0.0.12,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.27,<1",
    "aiosqlite>=0.20,<1",
    "ruff>=0.8,<1",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # Depends() in FastAPI signatures

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"
